Change metavar to refelct reality
Internally have 2 concepts of a stack:

* A stack instance specification (`InstanceSpec`) that may or may not
  (depending if you created or synced recently) correspond to a remote stack.
* The name of a remote stack instance you wish to interact with.

But for the user we should not care about this as this tool is ment to
reduce the conceptual diff between the instance specification and the
remote running stack.

Also the `_NAME` suffix in the metavar was redundant, as on CLI you can
only pass strings, you cannot pass "the full stack".